Hide cursors in unfocused windows immediately after switching focus
Currently changing window with <C-w>j/k results in cursors being drawn
in both the old window and the new window until an operation like
entering insert mode is made.

This commit tells vis to redraw the windows immediately after doing
window prev/next so that the cursors are updated in all windows and
hidden in the window that has lost focus.